pyproject.toml
[project]
name = "restful-with-fastapi-and-security-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "prisma (>=0.15.0,<0.16.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "redis (>=6.4.0,<7.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "authlib (>=1.3.0,<1.4.0)",
    "httpx (>=0.28.0,<0.29.0)",
    "python-multipart (>=0.0.9,<0.1.0)",
    "itsdangerous (>=2.2.0,<2.3.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
pytest-mock = "^3.15.0"

